Check for JavaScript
The JavaScript will pass a value to the server to let the server know
the script is present and working, meaning it’s already taken care of
updating the post-specific fields. If the user has JavaScript disabled
the value won’t be passed to the server, so the server will know it has
to handle updating the post-specific fields from the user profile.